pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-toolkit"
version = "0.1.0"
description = "A modular Python framework skeleton for RAG systems."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "pypdf>=5.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "python-dotenv>=1.0",
  "ruff>=0.4.0",
]
embedding = [
  "openai>=1.0",
]
zhipu = [
  "zai-sdk>=0.1",
]
llm = [
  "openai>=1.0",
  "zai-sdk>=0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
